Move some methods into audio file utilities
