+++ app/src/main/java/com/quran/labs/androidquran/dao/audio/AudioRequest.kt
@@ -3,6 +3,7 @@ package com.quran.labs.androidquran.dao.audio
 import android.os.Parcelable
 import com.quran.labs.androidquran.common.audio.model.QariItem
 import com.quran.data.model.SuraAyah
+import com.quran.labs.androidquran.common.audio.model.AudioPathInfo
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
@@ -13,7 +14,8 @@ data class AudioRequest(val start: SuraAyah,
                         val rangeRepeatInfo: Int = 0,
                         val enforceBounds: Boolean,
                         val shouldStream: Boolean,
-                        val audioPathInfo: AudioPathInfo) : Parcelable {
+                        val audioPathInfo: AudioPathInfo
+) : Parcelable {
   fun isGapless() = qari.isGapless
   fun needsIsti3athaAudio() =
       !isGapless() || audioPathInfo.gaplessDatabase?.contains("minshawi_murattal") ?: false
