+++ app/src/main/java/com/quran/labs/androidquran/util/AudioUtils.kt
@@ -8,12 +8,11 @@ import com.quran.data.model.SuraAyah
 import com.quran.data.model.audio.Qari
 import com.quran.labs.androidquran.common.audio.model.QariItem
 import com.quran.labs.androidquran.common.audio.util.QariUtil
-import com.quran.labs.androidquran.dao.audio.AudioPathInfo
 import com.quran.labs.androidquran.service.AudioService
+import timber.log.Timber
 import java.io.File
 import java.util.Locale
 import javax.inject.Inject
-import timber.log.Timber
 
 class AudioUtils @Inject constructor(
   private val quranInfo: QuranInfo,
@@ -76,22 +75,6 @@ class AudioUtils @Inject constructor(
     } else {
       "%03d%03d$AUDIO_EXTENSION"
     }
-  }
-
-  fun getLocalQariUrl(item: QariItem): String? {
-    val rootDirectory = quranFileUtils.audioFileDirectory()
-    return if (rootDirectory == null) null else rootDirectory + item.path
-  }
-
-  fun getQariDatabasePathIfGapless(item: QariItem): String? {
-    var databaseName = item.databaseName
-    if (databaseName != null) {
-      val path = getLocalQariUrl(item)
-      if (path != null) {
-        databaseName = path + File.separator + databaseName + DB_EXTENSION
-      }
-    }
-    return databaseName
   }
 
   fun getLastAyahToPlay(
@@ -281,25 +264,8 @@ class AudioUtils @Inject constructor(
     }
   }
 
-  fun getLocalAudioPathInfo(qari: QariItem): AudioPathInfo? {
-    val localPath = getLocalQariUrl(qari)
-    if (localPath != null) {
-      val databasePath = getQariDatabasePathIfGapless(qari)
-      val urlFormat = if (databasePath.isNullOrEmpty()) {
-        localPath + File.separator + "%d" + File.separator +
-            "%d" + AUDIO_EXTENSION
-      } else {
-        localPath + File.separator + "%03d" + AUDIO_EXTENSION
-      }
-      return AudioPathInfo(urlFormat, localPath, databasePath)
-    }
-    return null
-  }
-
   companion object {
     const val ZIP_EXTENSION = ".zip"
     const val AUDIO_EXTENSION = ".mp3"
-
-    private const val DB_EXTENSION = ".db"
   }
 }
