+++ common/audio/src/main/java/com/quran/labs/androidquran/common/audio/model/AudioPathInfo.kt
@@ -1,4 +1,4 @@
-package com.quran.labs.androidquran.dao.audio
+package com.quran.labs.androidquran.common.audio.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
